win and lose logic added!
user will now get feedback if they win or lose the game. you cannot "play new game" but i will add a button in there and will remove the grid or reinitialize it, and bring the form up once more.
also looking to add a flag functionality toggle, button that highlights it and will flag for one click